quang_huy_repo/requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pandas>=2.0
PyYAML>=6.0
SimpleITK>=2.3
scikit-learn>=1.3
